+++ tests/test_cli.py
@@ -9,11 +9,13 @@ from tests.unittestcase import UnitTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2024, Alex Laird"
-__version__ = "0.0.6"
+__version__ = "1.0.1"
 
 
 class TestCli(UnitTestCase):
     def setUp(self):
+        super().setUp()
+
         self.runner = CliRunner()
 
     def test_missing_credentials(self):