+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [1.0.1](https://github.com/alexdlaird/pyngrok/compare/1.0.0...1.0.1) - TBD
 ### Added
+- Auth flow now also checks session cookies in addition to parsing the page for signs of login.
 - Documentation improvements.
 - Test improvements.